# apps/web/lib/gen/model/transactionBase.ts
/**
 * Generated by orval v8.0.0-rc.3 🍺
 * Do not edit manually.
 * Core API
 * Core API
 * OpenAPI spec version: 0.0.1
 */
import type { Uid } from './uid';
import type { TransactionAmount } from './transactionAmount';

export interface TransactionBase {
  user_id: Uid;
  occurred_at: string;
  amount: TransactionAmount;
  /** @nullable */
  notes?: string | null;
  /** ID of recurring template if this transaction was created from one */
  recurring_template_id?: Uid;
}