# packages/openapi/api.yaml
openapi: 3.0.3
info:
  title: Core API
  description: |-
    Core API
  version: 0.0.1
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/transactions/create-expense:
    post:
      tags:
        - Transactions
      summary: Create expense transaction
      description: Create a new expense transaction in the transactions table
      operationId: createExpenseTransaction
      requestBody:
        required: true
        description: Expense transaction payload
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseTransactionPayload'
      responses:
        '201':
          description: Expense transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionExpense'
        '400':
          description: Error 400
          $ref: '#/components/responses/ErrorResponse400'
        '401':
          description: Error 401
          $ref: '#/components/responses/ErrorResponse401'
        '500':
          description: Error 500
          $ref: '#/components/responses/ErrorResponse500'
  /api/v1/transactions/create-income:
    post:
      tags:
        - Transactions
      summary: Create income transaction
      description: Create a new income transaction in the transactions table
      operationId: createIncomeTransaction
      requestBody:
        required: true
        description: Income transaction payload
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncomeTransactionPayload'
      responses:
        '201':
          description: Income transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionIncome'
        '400':
          description: Error 400
          $ref: '#/components/responses/ErrorResponse400'
        '401':
          description: Error 401
          $ref: '#/components/responses/ErrorResponse401'
        '500':
          description: Error 500
          $ref: '#/components/responses/ErrorResponse500'
  /api/v1/transactions/list:
    get:
      tags:
        - Transactions
      summary: List transactions
      description: List all transactions for the current user within a date range. This endpoint automatically materializes recurring transactions before returning the list.
      operationId: listTransactions
      parameters:
        - name: start_date
          in: query
          description: Start date for transaction range
          required: true
          schema:
            type: string
            format: date-time
            example: "2024-01-01T00:00:00Z"
        - name: end_date
          in: query
          description: End date for transaction range
          required: true
          schema:
            type: string
            format: date-time
            example: "2024-12-31T23:59:59Z"
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: '#/components/schemas/TransactionExpense'
                    - $ref: '#/components/schemas/TransactionIncome'
        '401':
          description: Error 401
          $ref: '#/components/responses/ErrorResponse401'
        '500':
          description: Error 500
          $ref: '#/components/responses/ErrorResponse500'
  /api/v1/transactions/update/{id}:
    patch:
      tags:
        - Transactions
      summary: Update transaction
      description: Update an existing transaction (partial update). Transaction type cannot be changed.
      operationId: updateTransaction
      parameters:
        - name: id
          in: path
          description: Transaction ID
          required: true
          schema:
            $ref: '#/components/schemas/UID'
      requestBody:
        required: true
        description: Transaction update payload
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTransactionPayload'
      responses:
        '200':
          description: Transaction updated
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/TransactionExpense'
                  - $ref: '#/components/schemas/TransactionIncome'
        '400':
          description: Error 400
          $ref: '#/components/responses/ErrorResponse400'
        '401':
          description: Error 401
          $ref: '#/components/responses/ErrorResponse401'
        '404':
          description: Transaction not found
          $ref: '#/components/responses/ErrorResponse404'
        '500':
          description: Error 500
          $ref: '#/components/responses/ErrorResponse500'
  /api/v1/transactions/delete/{id}:
    delete:
      tags:
        - Transactions
      summary: Delete transaction
      description: Delete an existing transaction
      operationId: deleteTransaction
      parameters:
        - name: id
          in: path
          description: Transaction ID
          required: true
          schema:
            $ref: '#/components/schemas/UID'
      responses:
        '204':
          description: Transaction deleted successfully
        '401':
          description: Error 401
          $ref: '#/components/responses/ErrorResponse401'
        '404':
          description: Transaction not found
          $ref: '#/components/responses/ErrorResponse404'
        '500':
          description: Error 500
          $ref: '#/components/responses/ErrorResponse500'

  # Recurring Templates
...
components:
  schemas:
    UID:
      type: string
      example: "12345abcde"
...
    TransactionAmount:
      type: number
      format: double
      example: 42.5
      description: Positive decimal amount
    TransactionBase:
      type: object
      properties:
        user_id:
          $ref: "#/components/schemas/UID"
        occurred_at:
          type: string
          format: date-time
          example: "2024-06-01T13:45:00Z"
        amount:
          $ref: "#/components/schemas/TransactionAmount"
        notes:
          type: string
          nullable: true
          example: "makeup"
        recurring_template_id:
          $ref: "#/components/schemas/UID"
          nullable: true
          description: "ID of recurring template if this transaction was created from one"
      required:
        - user_id
        - occurred_at
        - amount
    TransactionMeta:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UID"
        created_at:
          type: string
          format: date-time
          example: "2024-06-01T13:45:00Z"
      required:
        - id
        - created_at
    CreateExpenseTransactionPayload:
      allOf:
        - $ref: "#/components/schemas/TransactionBase"
        - type: object
          properties:
            type:
              type: string
              pattern: '^expense$'
              example: "expense"
            transaction_tag:
              type: string
              example: "want"
            expense_category_id:
              type: string
              example: "personal"
            expense_subcategory_id:
              type: string
              nullable: true
              example: "beauty_care"
          required:
            - type
            - transaction_tag
            - expense_category_id
    CreateIncomeTransactionPayload:
      allOf:
        - $ref: "#/components/schemas/TransactionBase"
        - type: object
          properties:
            type:
              type: string
              pattern: '^income$'
              example: "income"
            income_category_id:
              type: string
              example: "salary"
          required:
            - type
            - income_category_id
...
    UpdateExpenseTransactionPayload:
      type: object
      properties:
        type:
          type: string
          pattern: '^expense$'
          example: "expense"
        occurred_at:
          type: string
          format: date-time
          example: "2024-06-01T13:45:00Z"
        amount:
          $ref: "#/components/schemas/TransactionAmount"
        notes:
          type: string
          nullable: true
          example: "makeup"
        transaction_tag:
          type: string
          example: "want"
        expense_category_id:
          type: string
          example: "personal"
        expense_subcategory_id:
          type: string
          nullable: true
          example: "beauty_care"
      required:
        - type
    UpdateIncomeTransactionPayload:
      type: object
      properties:
        type:
          type: string
          pattern: '^income$'
          example: "income"
        occurred_at:
          type: string
          format: date-time
          example: "2024-06-01T13:45:00Z"
        amount:
          $ref: "#/components/schemas/TransactionAmount"
        notes:
          type: string
          nullable: true
          example: "salary payment"
        income_category_id:
          type: string
          example: "salary"
      required:
        - type
    UpdateTransactionPayload:
      oneOf:
        - $ref: "#/components/schemas/UpdateExpenseTransactionPayload"
        - $ref: "#/components/schemas/UpdateIncomeTransactionPayload"
      discriminator:
        propertyName: type
        mapping:
          expense: "#/components/schemas/UpdateExpenseTransactionPayload"
          income: "#/components/schemas/UpdateIncomeTransactionPayload"
    TransactionExpense:
      allOf:
        - $ref: "#/components/schemas/TransactionBase"
        - $ref: "#/components/schemas/TransactionMeta"
        - type: object
          properties:
            type:
              type: string
              pattern: '^expense$'
              example: "expense"
            transaction_tag:
              type: string
              example: "want"
            expense_category_id:
              type: string
              example: "personal"
            expense_subcategory_id:
              type: string
              nullable: true
              example: "beauty_care"
          required:
            - type
            - transaction_tag
            - expense_category_id
    TransactionIncome:
      allOf:
        - $ref: "#/components/schemas/TransactionBase"
        - $ref: "#/components/schemas/TransactionMeta"
        - type: object
          properties:
            type:
              type: string
              pattern: '^income$'
              example: "income"
            income_category_id:
              type: string
              example: "salary"
          required:
            - type
            - income_category_id
...
    GenericError:
      type: object
      description: Generic error response
      properties:
        status:
          type: integer
        error:
          type: string
          example: "INVALID_INPUT"
        message:
          type: string
          example: "Some error"
      required:
        - status
        - error
        - message
...
  responses:
...
    ErrorResponse400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericError'
    ErrorResponse401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericError'
    ErrorResponse404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericError'
    ErrorResponse500:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericError'